vars/buildImage: Configure parameters automatically in the jenkins UI
This will clear all the parameters set in the job's configuration
and will add the parameters defined here instead. This is done in
the first run of the generated job which means these parameters
will be available only from the second job run.

When launching the job, the user will be able to choose between
the default values provided by the library or edit and write
a different one.

Signed-off-by: Ana Guerrero Lopez <ana.guerrero@collabora.com>